allows manual backups/restores in backup/restore scripts
Summary:
The changes implemented in this diff allow the cloneCluster workflow to work... basically
it allows backups to backup a separate file called "metadata.json" which includes a list of dbs
that we want to backup which when we go to restore the backup later we can then access this list
through the "metadata.json" file.

Test Plan:
`make build-tools`
I also made sure this worked by loading the image and ensuring that the scripts worked how i envisioned
along with backingup and restoring dbs.

1st reviewer: @hdhanushkodi
2nd reviewer: @aluo, @tgummalla

Reviewers: hdhanushkodi, aluo, tgummalla, jkoh, davidgomes

Reviewed By: hdhanushkodi, davidgomes

Subscribers: davidgomes, tgummalla, hdhanushkodi, engineering-list, aluo

JIRA Issues: MCDB-2323

Differential Revision: https://grizzly.internal.memcompute.com/D41125