[deployment-docker] Fixes init.sql not being executed when root_password is set
Summary:
We recently introduced ROOT_PASSWORD as a mandatory environment variable for the cluster-in-a-box
Docker image. What we didn't realize is that adding this actually broke the automatic execution of
the `/init.sql` file.

Prior to introducing ROOT_PASSWORD, if there existed a file called `/init.sql` (in the root `/`
directory), it would get auto-executed when the container is initialized. This behavior is described
here:

https://github.com/memsql/deployment-docker#automatically-run-sql-when-cluster-in-a-box-initializes

This diff fixes the bug (which was reported by a user in https://github.com/memsql/deployment-docker/issues/2)
and it also changes our tests so that they actually test the `/init.sql` execution mechanism.

The fix is simply running the `init.sql` file _before_ changing the password. I tried a different fix
which was to run `init.sql` in the same place where it is being run now but that didn't work, and I
wasn't able to figure out why. However, I did note down what happened when I tried this if anyone
wants to take a look:

https://github.com/memsql/deployment-docker/issues/2#issuecomment-749671470

Test Plan:
```
make test-ciab
```

and

https://app.circleci.com/pipelines/github/memsql/deployment-docker/88/workflows/9365b5eb-1f15-41e0-8d83-437d97289015

Reviewers: carl, vtkachuk-ua

Subscribers: fmagalhaes, engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D44701