[deployment-docker] Fixes init.sql not being executed when root_password is set
Summary:
We recently introduced ROOT_PASSWORD as a mandatory environment variable for the cluster-in-a-box
Docker image. What we didn't realize is that adding this actually broke the automatic execution of
the `/init.sql` file.

Prior to introducing ROOT_PASSWORD, if there existed a file called `/init.sql` (in the root `/`
directory), it would get auto-executed when the container is initialized. This behavior is described
here:

https://github.com/memsql/deployment-docker#automatically-run-sql-when-cluster-in-a-box-initializes

More information here: https://github.com/memsql/deployment-docker/issues/2

Test Plan:
```
make test-ciab
```

and

https://app.circleci.com/pipelines/github/memsql/deployment-docker/88/workflows/9365b5eb-1f15-41e0-8d83-437d97289015

Reviewers: carl, vtkachuk-ua

Reviewed By: carl

Subscribers: fmagalhaes, engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D44701