Support graceful node termination
Summary:
Currently node images do not gracefully terminate on SIGTERM because
safe_memsqld is not a child process of /assets/startup-node so it does
not receive the SIGTERM sent to the container.

Fix this by trapping SIGTERM in /assets/startup-node and explicitly
sending it to the safe_memsqld pid.

Test Plan:
Added a test to test/node to verify that sending SIGTERM to the
container now correctly stops it.  Also manually verified that without
the startup script changes this new test fails.

Reviewers: vtkachuk-ua, mshcherbina-ua

Reviewed By: vtkachuk-ua

Subscribers: tlucas, rodrigo, engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D56243