Changing the memsql user uid and gid
Summary:
In the old centos base image, the memsql user was create simply with the dockerfile `USER memsql`
command. This resulted in a memsql user with uid 999 and gid 998. When we bumped the base image to
alma, there already was a uid 999 for the user systemd-coredump and a gid 998 for the group
rancher.

As a result, we get into a weird state when we bump the memsql image from centos to alma because
the files on the persistent volume are owned by uid 999.

To fix this, this change adds a script to move make the memsql user the same as it was previously
in the centos image.

Test Plan: make test

Reviewers: rodrigo, vtkachuk-ua, sasha, mwilder

Reviewed By: rodrigo, mwilder

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D56598