[tools] Re-add the "backup" script that was removed on D55785
Summary:
On https://grizzly.internal.memcompute.com/D55785 we moved some scripts to a separate image that is only used by the managed-service, including the `backup/memsql_backup.py`.

But now we realized that this script is being used by the self-managed operator so we need to re-add it again to the `memsql/tools` image, see https://singlestore.zendesk.com/agent/tickets/21763.

Test Plan: N/A

Reviewers: vtkachuk-ua, noprysk-ua

Reviewed By: vtkachuk-ua, noprysk-ua

Subscribers: engineering-list, davidgomes

JIRA Issues: PLAT-6227

Differential Revision: https://grizzly.internal.memcompute.com/D56917