Skipping recovery check in waitStart
Summary:
When a leaf noode boots up with a disk from an previous instance, if the
root password has changed, the `waitStart` function will get permenantly blocked.
The problem is that the databases are still in the pending state and the recovery state from
memsqlctl is `Recovering`. As a result we do not proceed to update the root password.

This situation results in a deadlock with the k8s operator,
1) To resolve the pending state we need to add the leaf on the MA
2) To attach the leaf we need to be able to connect using the root pw
3) We dont have the root password and are in the pending state

Test Plan:
ci
https://app.circleci.com/pipelines/github/memsql/deployment-docker/443/workflows/bbb5ebde-cd78-41cb-ae0a-931f2e7057ca

Reviewers: mwilder, adam, rodrigo

Reviewed By: rodrigo

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D59174