Checking for dead pid in waitStart
Summary:
If you set a bad memsql config, memsqld will exit. As a result we end up in an infinite loop,
memsqld is gone but we are waiting for `memsqlctl describe-node --property IsConnectable`

because this script is the main process of the container, the container will just sit there and not
exit.

to fix this, we should just check the memsqld pid still exists

Test Plan: ci

Reviewers: mwilder, adam, rodrigo

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D59368