run publish gcr on condition
Summary: Images that are not node images do not need to be published to GCR singlestore public and do not have the makefiles, causing the job to fail. This MR adds a conditional step that only runs this publish job if the publish job starts with "node".

Test Plan: TODO

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D59819