Add 7.8 to circleci config.
Summary:
We missed doing this.  Alex did part of it when adding 8.0 here (https://grizzly.internal.memcompute.com/D60403), but forgot the .yml file change so we can't relase 7.8 node images.
When we released 7.8.20 we didn't add it either (https://grizzly.internal.memcompute.com/D60521) so we don't have a 7.8.20 node imaged published right now.

I'm following the wiki on adding this (it said to do the same thing as this diff): https://grizzly.internal.memcompute.com/D50082 which add 7.3.

I'm not 100% sure how to kick off circle again for 7.8 so we can release it.

**Design doc/spec**:
**Docs impact**: none

Test Plan: none

Reviewers: davidgomes, mgiannakopoulos, ymolinar, asteere

Reviewed By: davidgomes

Subscribers: mojdeh, engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D60566