Start the exporter with sdb-admin in ciab image
Summary:
Current exporter start method conflicts with user setting up monitoring on their cluster through Toolbox. In addition, the exporter currently doesn't even work because the root password, which it needs to query the cluster, is not passed to it.

After this change, the exporter is still started on the same port as before, but actually works and is suitable for further monitoring setup

Test Plan:
`make test` in root dir
manually tested that the exporter is started and accessible

Reviewers: vtkachuk-ua

Reviewed By: vtkachuk-ua

Subscribers: engineering-list

JIRA Issues: PLAT-6482

Differential Revision: https://grizzly.internal.memcompute.com/D61126