removing non-gcr redhat jobs
Summary:
we havent pushed to redhat marketplace in over 2 years, all the scan.connect.redhat.com
functionality has been totally deprecated. Now they require you to run this tool and upload results
https://github.com/redhat-openshift-ecosystem/openshift-preflight
its an absoulte pain when I tried to do this with the operator.

If we plan on actually supporting the marketplace going forward we need to jumo through all their
hoops.

This diff removes all the non-functional code to push to redhat but leaves the gcr logic

Test Plan: ci

Reviewers: adam, rodrigo, mwilder

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D62264